refactor(models): raname player_hand to hand. rename PokerHandCalculator to HandCalculator and rename calculate_hand_ranks to calculate_all_possible_hand_ranks.
remove player from player_hand because it is redundant(i.e. participants are all players, there isn't x_Hand, where x is something that is not a player). rename calculate_hand_ranks to make it more explicit that it is all possible hand ranks(i.e. it could have been a subset)